merged most of the branches and fixed some bugs
added a build script in package.json which makes a uploadTemplate.js so you can create a templated upload.

many style changes allowing for the crops div to have individual crops. I coudn't figure out how to add a favorite button to each of them without giving each of them their own distinction.

changed the add.evenEventListener in index.handlebars. I don't know if I was supposed to mess with this but on my end so far I am able to crop images on my end. Please tell me if this doesn't work

added two scripts to main.handlebars in order to allow the uploadTemplate.js to work.

Changed how upload.handlebars worked. it now takes a boolean named all. if the function in event.handlebars gives a boolean that is true the upload.handlebars will now add a button that has been styled to be on the bottom right of uploaded image and will now give the image id = 'crop' instead of 'image'.
added the base call to request /favorites tab that will need to be implemented.